requirements.txt
pandas~=1.5.1
PyYAML~=6.0
dataclasses~=0.6
sklearn~=0.0
scikit-learn~=1.1.3
numpy~=1.23.4
click~=8.1.3